examples/SliceServer/requirements.txt
numpy
openvds
Flask
Pillow
